replace var for todoTitle with const, replace event.target.title.value = '' with event.target.title.value.reset, create new folder called components and move working components to it
