+++ src/App.jsx
@@ -1,8 +1,7 @@
 import React from "react";
-import TodoList from "./TodoList";
-import AddTodoForm from "./AddTodoForm";
-import "./App.css";
-
+import "./style/App.css";
+import TodoList from "./components/TodoList";
+import AddTodoForm from "./components/AddTodoForm";
 
 function App() {
