+++ src/components/TodoList.jsx
@@ -12,7 +12,7 @@ function TodoList() {
         <>
             <ul>
                 {todoList.map((item) => (
-                    <TodoListItem key={item.id} todo={item.title}/>
+                    <TodoListItem key={item.id} todo={item.title} />
                 ))}
             </ul>
         </>
